- creating crud - react redux
