+++ src/components/counter.js
