+++ src/store/index.js
@@ -4,10 +4,13 @@ const initialiState = {
   count: 0
 }
 
+
+
 const reducer = (state = initialiState, action) => {
   console.log('reducer running', action);
   switch (action.type) {
     case 'INCREMENT':
+                                       //new count is going to be
       return Object.assign({}, state, { count: state.count + 1 });
     case 'DECREMENT':
       return Object.assign({}, state, { count: state.count - 1 });
@@ -16,6 +19,8 @@ const reducer = (state = initialiState, action) => {
   }
 }
 
-const store = createStore(reducer);
+const store = createStore(reducer, 
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  );
 
 export default store;
